requirements.txt

fastapi
uvicorn
pydantic
openai>=1.0.0
moviepy
pillow
ffmpeg